+++ src/CleanRegex/Internal/Replace/Counting/AtLeastCountingStrategy.php
@@ -2,6 +2,7 @@
 namespace TRegx\CleanRegex\Internal\Replace\Counting;
 
 use TRegx\CleanRegex\Exception\ReplacementExpectationFailedException;
+use TRegx\CleanRegex\Internal\Model\GroupAware;
 
 class AtLeastCountingStrategy implements CountingStrategy
 {
@@ -16,7 +17,7 @@ class AtLeastCountingStrategy implements CountingStrategy
         $this->limitPhrase = $phrase;
     }
 
-    public function count(int $replaced): void
+    public function count(int $replaced, GroupAware $groupAware): void
     {
         if ($replaced < $this->limit) {
             throw ReplacementExpectationFailedException::insufficient($replaced, $this->limit, $this->limitPhrase);