+++ src/CleanRegex/Internal/Replace/Counting/CountingStrategy.php
@@ -1,7 +1,9 @@
 <?php
 namespace TRegx\CleanRegex\Internal\Replace\Counting;
 
+use TRegx\CleanRegex\Internal\Model\GroupAware;
+
 interface CountingStrategy
 {
-    public function count(int $replaced): void;
+    public function count(int $replaced, GroupAware $groupAware): void;
 }