+++ src/CleanRegex/Internal/Replace/Counting/IgnoreCounting.php
@@ -1,9 +1,11 @@
 <?php
 namespace TRegx\CleanRegex\Internal\Replace\Counting;
 
+use TRegx\CleanRegex\Internal\Model\GroupAware;
+
 class IgnoreCounting implements CountingStrategy
 {
-    public function count(int $replaced): void
+    public function count(int $replaced, GroupAware $groupAware): void
     {
     }
 }